Refactoring to comply with DDD and Hexagonal architecture standards
