+++ src/main/java/com/store/price_fetcher/application/ports/outbounds/PriceRepository.java
@@ -1,9 +1,9 @@
-package com.store.price_fetcher.domain.repositories;
+package com.store.price_fetcher.application.ports.outbounds;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-import com.store.price_fetcher.domain.model.Price;
+import com.store.price_fetcher.domain.entities.Price;
 
 public interface PriceRepository {
     Optional<Price> findPrice(int productId, int brandId, LocalDateTime dateTime);
